+++ test/analysis/interpolation/multi_linear_interpolation_test.dart
@@ -15,6 +15,38 @@ void tests() {
     expect(() => ir.valueAt(-1), throwsArgumentError);
     expect(() => ir.valueAt(10), throwsArgumentError);
   });
+  test('Multi linear interpolator, using infinity', () {
+    var ir = MultiLinearInterpolator([0, 15, double.infinity], [15, 0, 0],
+        extrapolate: true);
+    expect(ir.valueAt(0), 15);
+    expect(ir.valueAt(-10), 25);
+    expect(ir.valueAt(15), 0);
+    expect(ir.valueAt(20), 0);
+    expect(ir.valueAt(double.infinity), 0);
+  });
+  test('Multi linear interpolator, using +/-infinity', () {
+    var ir = MultiLinearInterpolator(
+        [double.negativeInfinity, 0, 15, double.infinity], [15, 15, 0, 0],
+        extrapolate: true);
+    expect(ir.valueAt(0), 15);
+    expect(ir.valueAt(-10), 15);
+    expect(ir.valueAt(15), 0);
+    expect(ir.valueAt(20), 0);
+    expect(ir.valueAt(double.infinity), 0);
+    expect(ir.valueAt(double.negativeInfinity), 15);
+  });
+
+  test('Multi linear interpolator, discontinuous function (theta function)',
+      () {
+    // discontinuity needs to be explicitly specified
+    var ir = MultiLinearInterpolator(
+        [double.negativeInfinity, 0, double.minPositive, double.infinity],
+        [0, 0, 1, 1]);
+    expect(ir.valueAt(0), 0);
+    expect(ir.valueAt(0.5), 1);
+    expect(ir.valueAt(-1), 0);
+  });
+
   test('Multi linear interpolator, 2 points only', () {
     var ir = MultiLinearInterpolator([0, 1], [0, 4], extrapolate: true);
     expect(ir.valueAt(0), 0);